src/web/requirements.txt
django
django-bootstrap4
djangorestframework
djangorestframework_simplejwt